[UPD] Break the form out of the contact section component
